Compress graphics data
All graphics data is 1bpp.

... 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, ...

First it is grouped into alternating runs of
0 and 1 bits, with a max length of 15 (i.e. 4 bits).

... 1, 1, 1, 5, 10, 2, 3, 5, 10, 10, 10, 10, 10, 10, 10, 8...

Then these 4-bit values are scanned to find repeated patterns in
previous output (like LZ77 compression), with a minimum length of at
least 6 values. The back-references are displayed as `(dist, len)`
below.

... 1, 1, 1, 5, 10, 2, 3, 5, 10, (-1, 6), 8...

Then groups of literal values and back-references are alternated, where
runs of literals are prefixed by a count (9 below).

... [#9, 1, 1, 1, 5, 10, 2, 3, 5, 10], (-1, 6), [#N, 8, ...]...

Finally each value is written with a fixed bit length (using huffman
encoding was explored, but did not produce significantly better
compression, and would require more code to decompress). The literal
count is 7 bits, the literal values are 4 bits, the back-reference
distance is 7 bits, and the back-reference length is 7 bits. The
distance is always negative, so is encoded as a positive number.

... [0001001, 0001, 0001, 0001, 0101, 1010, 0010, 0011, 0101, 1010],
(0000001, 0000110), ...

The data is decompressed to 8bpp pixels (0=transparent, 1=black), which
allows some simplifications of the `$blit` function.