Draw dino on top of clouds
This was kinda tricky to make work. Since the drawing is used for
collision detection, we only want to draw obstacles before we draw the
dino (this is the way it worked previously). But this also means that
we'll draw the ground and clouds on top of the dino. The ground is hard
to notice (though we would notice it if the dino had an outline). The
clouds were pretty easy to see, however.

The solution here is to only check for a specific color when determining
whether a collision occurred. We can set the obstacles to a slightly
different color than other objects, and then we can draw in any order,
as long as the dino is still after the obstacles. Some other size
optimizations rely on the dino+obstacles being lower in memory, so we
draw the objects in reverse order now.