Step 5: Get mouse input, move colored pixel
This introduces `i32.load8_u`, which loads a byte from memory.

In JavaScript, we handle the `mouseup`, `mousedown`, and `mousemove`
events by writing the current mouse location and mouse buttons to Wasm
memory.

Byte 0 is the x-coordinate, byte 1 is the y-coordinate. Byte 2 is a
bitfield of mouse-buttons, where each bit is 1 if that mouse button is
pressed. The x and y coordinates are clamped to the range [0, 150).

To test this behavior, we take the current (x, y) mouse coordinate and
draw a red or blue pixel; red if any button is pressed, or blue
otherwise.