Step 7: Drawing sprites
We replace the `$fill-rect` function with `$draw-sprite`. The difference
is that `$fill-rect` uses a single color, and `$draw-sprite` reads the
color from memory.

The source pixel is indexed by `(w * j + i) * 4`, which is similar to
the destination index, but uses the sprite width (`w`) instead of the
screen width (`150`).

The equivalent JavaScript would be:

```js
let mem = new DataView(buffer);

function drawSprite(x, y, w, h, src) {
  for (let j = 0; j < h; ++j) {
    for (let i = 0; i < w; ++i) {
      putPixel(x + i, y + j, mem.getUint32(src + (w * j + i) * 4, true));
    }
  }
}
```

The sprite is defined below in the `data` section using a sequence of
bytes encoded as "\XX" where XX is the hexadecimal value of the byte.