Step 8: Mask out alpha pixels when drawing sprites
The previous commit would always draw each color from the sprite over
the color that already was in the destination buffer. That would be fine
if the sprites were always square, but since the sprite is round, some
pixels should by transparent.

This commit adds transparency by treating the color `0` as a completely
transparent pixel, so it doesn't draw over the pixel.

This is done by using the `if` instruction, which will only execute the
contained instructions if the conditional expression is non-zero.

The equivalent JavaScript would be:

```js
let mem = new DataView(buffer);

function drawSprite(x, y, w, h, color) {
  for (let j = 0; j < h; ++j) {
    for (let i = 0; i < w; ++i) {
      let pixel = mem.getUint32((w * j + i) * 4, true);
      if (pixel) {
        putPixel(x + i, w + j, pixel);
      }
    }
  }
}
```