Step 11: Drawing a scaled sprite
This step shows the first example of using the `f32` type, which is a
32-bit floating point type.

The `$draw-sprite` function is modified to take two more parameters,
`$dw` and `$dh`, the destination width and height. The `$w` and `$h`
parameters are now renamed `$sw` and `$sh`, the source width and height.

The sprite is drawn with a size of `$dw x $dh` pixels, either scaled up
or scaled down. The scaling width factor is `$sw / $dw`, and the scaling
height factor is `$sh / $dh`. The source pixel can be determined by
multiplying the destination x- and y-coordinate by this factor. This
would be expressed in JavaScript as:

```js
let mem = new DataView(buffer);

function drawSprite(x, y, src, sw, sh, dw, dh) {
  let dx = sw / dw;
  let dy = sh / dh;
  for (let j = 0; j < dh; ++j) {
    for (let i = 0; i < dw; ++i) {
      let pixel = mem.getUint32(src + sw * j * dy + i * dx, true);
      if (pixel != 0) {
        putPixel(x + i, y + j, pixel);
      }
    }
  }
}
```

Note that type conversions must be explicit in WebAssembly, so the
conversion from `f32` to `i32` uses the `i32.trunc_f32_s` instruction,
and the conversion from `i32` to `f32` uses the `f32.convert_i32_s`
instruction, since these values will be interpreted as signed.