Step 12: Draw a 2d grid of sprites
To draw an 8x8 grid, we need to have an 8x8 array of sprites to draw. We
could use one byte per cell, but it will be more convenient later to use
8 grids of 1 bit per cell.

In this step, we implement drawing a single 8x8 grid, where each cell is
drawn using the same sprite, or not at all.

Since the grid may be sparsely populated, and is exactly 64-bits, we can
iterate over the cells in a clever way: by succesively finding and
clearing the `1` bits in the grid. For example, if the grid value is the
following binary sequence:

    0000010000001000010000001100000010000000010000000000000000000100

Then it will be drawn as the following grid, where the least significant
bit is in the lower-left corner, and the most signficant bit is in the
upper-right corner:

    00100000
    00010000
    00000010
    00000011
    00000001
    00000010
    00000000
    00100000

To find the index of the lowest `1` bit in a 64-bit integer, we can use
`i64.ctz` ("Count Trailing Zeroes"). In our example, this would give a
value of `2`. We then convert the index into an x and y coordinate using
the following formulas:

    horizontalPadding = 7
    cellWidthWithSpacing = 17
    screenWidth = 150
    cellRows = 8

    cellX = idx % cellRows
    cellY = idx / cellRows
    pixelX = horizontalPadding + cellX * cellWidthWithSpacing
    pixelY = (screenWidth - cellWidthWithSpacing - horizontalPadding) -
               cellY * cellWidthWithSpacing

We draw a sprite at this location, then clear the lowest bit, and repeat
until the grid is empty. To clear the lowest bit, we can use the
following technique:

    x = x & (x - 1)

Note that the WebAssembly code here uses another common trick; if you
are calculating the quotient and remainder and the divisor is a power of
two, you can use the "and" and "shift" operators instead:

    q = n / 8    r = n % 8

is equivalent to:

    q = n >> 3   r = n & 7