Step 9: clip pixels that are out of bounds
When drawing pixels, we want to make sure we properly handle
out-of-bounds pixels. Otherwise we could improperly wrap at the
left/right edges, or clobber other memory if at the top/bottom edges.

To fix it, we can check the x- and y-coordinates in `$put-pixel`. We use
`br_if 0`, which is the same as writing `if (...) return;`, since the
zeroeth label at this point is the function's label.

This commit also introduces `i32.ge_u` (greater-than or equal,
unsigned), and `i32.or`. The `i32.or` instruction is a bitwise or, but
it can work as a logical or since the result of the comparison operators
is always `0` or `1`.

The equivalent JavaScript would be:

```js
let mem = new DataView(buffer);

function putPixel(x, y, color) {
  if ((x >= 0 && x < 150) || (x >= 0 && y < 150)) return;
  mem.setUint32((y * 150 + x) * 4, color, true);
}
```

Note that we don't have to do the `x >= 0` comparison in Wasm, since we
are doing an unsigned comparison, so it will always be >= 0. (It's
possible to do this in JavaScript too, using the >>> operator:
`((x >>> 0) < 150)` )