Step 15: Animate x,y,w,h values
This step adds animation to the grid cells.

To do this, we first give each cell an additional 4 values: the
x-offset, y-offset, width-offset, and height-offset. If these values are
all 0, then the cell will be drawn at its normal location and at its
normal size.

These values are treated as signed, so if the x-offset is -4, then the
cell will be drawn 4 pixels to the left. If the width-offset is 7, then
the cell will be drawn 7 pixels wider than the base size (resulting in a
width of 16 + 7 = 23 pixels).

This change is done by using the `i32.load8_s` instruction in the
`$draw-cell` function, which loads an 8-bit signed value, and extends it
to 32-bits.

These offsets are then animated. For each cell, and for each offset
(x,y,width,height), there is a starting value and an ending value. For
each cell there is also a "time" value, which is in the range [0, 1].
If the time is 0, then each offset for that cell is the starting value,
and if the time is 1 then each offset for that cell is the ending value.

If the time is between 0 and 1, then the offset is is interpolated.
Instead of using linear interpolation (which doesn't look very nice), we
an "ease-out cubic" function: `t * (3 + t * (t - 3))`. It is called
"ease-out" because it starts moving quickly, but as the time approaches
1, it slows down or "eases" into the ending value.

Note the way the animation values are stored in memory. Using C syntax:

    struct Offsets {
      s8 x, y, w, h;
    };

    0x3200: struct Offsets current[64];  // current offsets
    0x3300: struct Offsets start[64];    // start offsets
    0x3400: struct Offsets end[64];      // ending offsets
    0x3500: f32     time[64];            // time in range [0, 1]

This is unlike how you might typically write a structure like this,
where the animation values are grouped:

    struct AnimatedValue {
      s8 current, start, end;
    };

    struct Cell {
      struct AnimatedValue x, y, w, h;
      f32 time;
    };

    struct Cell cells[64];

This is sometimes "Array Of Structures" or "AoS". The data layout used
here instead is sometimes called "Structure of Arrays" or "SoA".

The main benefit for us here is that we can group all starting values,
all current values, etc. This means that the `$animate` function can
treat them identically; that is, we can iterate over all the `start`
values in one big block, without worrying whether they are x-offsets,
y-offsets, width-offsets, or height-offsets.

The `$animation` function increments all the time values for each cell
(capping it at 1), and then interpolates each value.

Conveniently, the time value is 4 bytes and each of the 4 cell offsets
is 1 byte, so the address of the time value can be calculated by masking
off the low 2 bits of the current offset. i.e. if we view the first 4
cells:

  address   0123456789abcdef...
  cell      0   1   2   3   ...
  current   xywhxywhxywhxywh...
  start     xywhxywhxywhxywh...
  end       xywhxywhxywhxywh...
  time      t   t   t   t   ...