Step 14: Get grid mouse location
In this step we calculate the grid position of the mouse, and represent
it as a 64-bit value. The value will have at most one bit set, at the
grid index of the mouse. For example, if the mouse is in the lower-left
corner, then the return value will be the binary sequence:

    0000000000000000000000000000000000000000000000000000000000000001

Since the least-significant bit of the 64-bit grid value represents the
lower left corner.

If the mouse is not in the grid (e.g. it is in the border surrounding
the grid), then the result of the function will be `0`.

To calculate the index, we perform the inverse of the calculation used
to find the pixel location:

    horizontalPadding = 7
    cellWidthWithSpacing = 17
    screenWidth = 150
    cellRows = 8

    offsetX = pixelX - horizontalPadding
    offsetY = (screenWidth - horizontalPadding - 1) - pixelY

    index = (pixelY / cellWidthWithSpacing) * cellRows +
            (pixelX / cellWidthWithSpacing)
    bit = 1 << index

This introduces some new instructions: `i64.extend_i32_u` converts from
an unsigned `i32` to an `i64`; and `i32.div_s` performs signed division
on to `i32` values.