Step 17: Animate cells as we hover over them
In this step, we animate the currently hovered cell by making it grow
larger. We only want to start this animation when the cell is newly
hovered, and we want to shrink it back to its original size when it is
no longer being hovered.

To do so, we keep track of the previous mouse location, using a
WebAssembly global. Unlike WebAssembly memory, a global has a type, so
it can be accessed more like a local variable (using `global.get` and
`global.set`).

If the current mouse position is not equal to the last mouse position,
we can start the animation. The function `$animate-cell` takes a grid
"bit" to animate and sets the desired x, y, width, and height ending
offsets.

Since we know the exact values we want to set (either {-4,-4,8,8} to
scale up or {0,0,0,0} to reset) we can save some space by writing these
values as a 32-bit value instead of writing 4 8-bit values. So
`$animate-cell` takes a single `i32` parameter. Since WebAssembly is
little-endian, the number is written in the reverse order that the bytes
are written in text, so the parameter is called `$h_w_y_x` to make this
clear.

To start an animation, the current offset is copied to the starting
offset (so the animation begins at the current position). The ending
offset is the function parameter. And the time value is set to one minus
the current time. This way if the cell is currently animating, it will
not completely reset its animation.