Step 19: Implement clicking/dragging on a cell
In this step we start implementing user interaction on the grid. When
the user clicks on a cell, they can drag the cell around. When they
release the mouse button it returns to its original location.

To implement this, we keep track of the current game state. Initially
the game is in the "idle" state (0). When a cell is clicked the game
moves to the "mouse-down" state (1). When the mouse is released it
returns to the "idle" state.

The game state is stored as an `i32` global. The `$run` function
switches on the current state using the `br_table` instruction, which
branches to a label based on a given `i32` value. Branching to a `block`
label goes to its end, so the structure of the switch is:

    block $label0
    block $label1
    block $label_default
      br_table $label0 $label1 $label_default (global.get $state)

    end $label0
      ;; branch here if state is 0

    end $label1
      ;; branch here if state is 1

    end $label_default
      ;; branch here if state is not 0 or 1

Note that we can provide a label at the top and bottom of the `block` to
make it clearer.

In the "idle" state, we perform the hovering animation like before, but
also move to the "mouse down" state if the mouse button is clicked. We
need to remember where the mouse was clicked, so we copy the current
mouse position to the two bytes at address 3 and 4.

In the "mouse down" state, we update the active cell so that its ending
position is the difference between the current mouse position and where
we originally clicked. This way it will look like we're dragging it
around. Finally, if the mouse button is released then we go back to the
"idle" state.