Step 20: Clamp dragging to 4 adjacent grid cells
The previous commit has preliminary dragging behavior, but we want to
only allow the user to drag to one of the 4 neighboring cells, above,
below, left or right.

We calculate the difference between the current mouse position and the
original click position. If the difference in x coordinate is greater,
then we treat it as a horizontal drag, otherwise it is a vertical drag.
This is done by comparing the absolute values of differences, using the
`f32.abs` instruction.

For a horizontal drag, we zero out the vertical motion, and likewise we
zero out the horizontal motion for a for a vertical drag. For the drag
direction we want to clamp the motion to 17 pixels in either direction
(where 17 is the distance between adjacent cells).

We clamp the distance using `f32.min` of the absolute value, then use
`f32.copysign` to copy the original sign back into the value.

For example, if the difference is (-18, 3), then the absolute value of
the x-coordinate (18) is greater than the absolute value of the
y-coordinate (3). Since this is a horizontal motion, we zero out the
vertical component. Then we choose the minimum of 18 and 17 (which is
17), and copy the original sign to produce the final offset (-17, 0).