Step 21: Swapping animation when dragging
The current dragging animation is clamped to the nearest 4 neighbors,
but doesn't show that the neighboring cell will be swapped. To indicate
this, as we move the mouse toward another cell we want the other cell to
move toward this one.

We first get the position of the mouse that we would swap with (i.e. one
of of the 4 neighboring cells). If the mouse is still over the current
cell, we don't know which one to swap with, so we do nothing. Otherwise
we set the adjacent cell's x and y offsets to the negation of this
cell's x and y offsets. This works because the offsets are all relative
to the cell's starting location.

We need another global too: `$click-mouse-bit` is used to store the
position of the currently clicked cell. `$prev-mouse-bit` stores the
location of the last hovered cell. Previously we used `$prev-mouse-bit`
to keep track of the currently clicked cell, but now we need to use
`$prev-mouse-bit` to reset the animation on a cell that we decided not
to swap with.