Step 13: Implement $draw-grids
This step adds the `$draw-grids` function, which draws multiple grids by
calling `$draw-grid`.

The `i64.load` instruction is used to read each of the 8 grids and pass
them to `$draw-grid`. Each grid represents an 8x8 grid of a single
image, so all 8 grids must be drawn to be sure that all cells are
filled.

The grids are stored at memory address 0x3000, and each is 8 bytes. So
the addresses are as follows:

    0x3000: cells 0
    0x3008: cells 1
    0x3010: cells 2
    0x3018: cells 3
    0x3020: cells 4
    0x3028: cells 5
    0x3030: cells 6
    0x3038: cells 7

Given this representation, it's possible to draw multiple images in the
same location (if there is a `1` in the same bit of multiple grids), but
we'll make sure that doesn't happen.

Since each sprite is now 256 bytes (16x16x1 byte per pixel), we can
caluclate the source address of each sprite using the formula:

    src = 0x100 + (i * 256)

Where the first sprite is at address 0x100. Instead of multiplying by
256, however, we use a left-shift of 8 bits, which is equivalent.