Step 18: Update $draw-grids to use mask
The previous step added a hover effect, but sometimes the hovered cell
appears behind the cells below it. This happens when the cells below are
drawn after the hovered cell.

To fix this, we can draw the hovered cell again, after we've finished
drawing all the cells.

This can be done without much extra code by passing a "mask" to the
`$draw-grids` function. This mask specifies which grid cells to draw.
The first time, we draw all cells. The second time we only draw the
hovered cell. To implement this, we use `i64.and` to mask away the grid
bits when loading from memory.