Step 24: Swap grid cells when drag completes
The next step to implementing the game is to swap adjacent cells when
the mouse is released. We first check that the cell is valid (we can't
swap with a cell past the edge of the grid), and that we aren't trying
to swap with ourselves.

If we can swap, then we iterate through all the grids and swap the bits
at each cell's indexes.

So for example, if we want to swap the cells A and B here:

   ........
   ........
   ....A...
   ....B...
   ........
   ........
   ........
   ........

Then we need to swap the bits at those indexes in each of the 8 grids,
leaving all the other bits the same. So if A is sprite 3 and B is sprite
5:

                               A       B
   cell 0: ....................0.......0...................................
   cell 1: ....................0.......0...................................
   cell 2: ....................0.......0...................................
   cell 3: ....................1.......0...................................
   cell 4: ....................0.......0...................................
   cell 5: ....................0.......1...................................
   cell 6: ....................0.......0...................................
   cell 7: ....................0.......0...................................

Then after we swap the bits in those colums we'll have:

                               A       B
   cell 0: ....................0.......0...................................
   cell 1: ....................0.......0...................................
   cell 2: ....................0.......0...................................
   cell 3: ....................0.......1...................................
   cell 4: ....................0.......0...................................
   cell 5: ....................1.......0...................................
   cell 6: ....................0.......0...................................
   cell 7: ....................0.......0...................................

To swap bits, we can create a mask with these two columns:

           0000000000000000000010000000100000000000000000000000000000000000

If we mask the grid and the result is 0, then neither of the bits are
set, so we don't need to swap anything. Similarly if the result is equal
to the mask, then both of the bits are set, so we don't need to swap
anything either. Otherwise we know that one bit is set and the other
isn't. To swap them, we can xor the grid with our mask, e.g.

           ....................0.......1...................................
    xor    0000000000000000000010000000100000000000000000000000000000000000
     ==    ....................1.......0...................................