Step 16: Use start function
This step introduces the "start" function, which is called when the
WebAssembly module is instantiated.

This change is only a size optimization, reducing the module from 3125
bytes to 2897 bytes. Rather than specifying each of the ending offsets
in a data section, we can write them into memory when the module is
first instantiated.