Step 25: Clear matches
After we swap cells, we want to check whether this swap created a match
of 3 of the same cell horizontally or vertically. This is where using a
bitfield for the grid is extremely convenient.

To check for a match we can look for a pattern of 3-in-a-row or
3-in-a-column by using a mask. We start with the mask in the lower left
corner of the grid, and shift it by one bit, to find all possible
locations. If `(grid & mask) == mask`, then the pattern has matched. The
only problem is that some shifted patterns are not valid. For example,
the horizontal pattern:

     ........
     ........
     ........
     ........
     ........
     ........
     ........
     111.....

Shifted by 7 yields:

     ........
     ........
     ........
     ........
     ........
     ........
     11......
     .......1

Which is not a valid pattern. So for each potential shift, we need a
separate value to determine whether it is valid. Since there are 64
shifts, we can store this in a 64-bit integer. For the horizontal
pattern, the following shifts are valid:

     111111..
     111111..
     111111..
     111111..
     111111..
     111111..
     111111..
     111111..

The `$match-pattern` function handles this, given a `$grid`, a
`$pattern`, and `$shifts` will produce a result of all cells where this
pattern matches. This function loops over all the shift amounts and
stops when it is 0 (no more possible valid shifts).

This only checks one pattern against one grid, however. So the
`$match-patterns` function loops over both patterns (horizontal and
vertical) for one grid and accumulates the matches for either using
`i32.or` to combine them.

Finally, `$match-all-grids-patterns` loops over all the grids and looks
for all patterns that match, combining them all into one 64-bit value.

When a pattern matches, the behavior of the game is to remove those
cells. The `$clear-pattern` function does this by looping over all the
grids and zeroing out those bits (using `grid & ~mask`).