Step 26: If no match, reset swap
In a match3 game, you're only allowed to swap two cells if they produce
at least one match. So if there is no match, then we should swap them
back.

We do that by checking to see whether the result of
`$match-all-grids-patterns` (which is forwarded through
`$clear-pattern`) is non-zero; if it is zero there were no cells
matched. In that case, we swap the cells again and animate them back to
their original position.