Step 27: Move cells down after clearing
The rules of match3 are that after cells are removed, the cells above
fall down to fill in their place. If there are no cells above, then new
cells are randomly created to fill in the holes.

The `$move-down` function handles this behavior. It is passed a grid of
all of the "empty" cells (i.e. the recently cleared cells).

Like `$draw-grid`, it loops over all the bits, starting with the
least-significant `1` bit, finding the index of that bit with `i64.ctz`.
We can then find the next `1` bit above this one, using the binary mask,
which is shifted to the current index:

  100000001000000010000000100000001000000010000000100000001

If there is a bit found above, then we move it down by swapping the
current index and the bit we found (using the same
`$swap-all-grids-bits` as used before).

If there isn't a bit found above, then we need to fill it in with a
random cell. We import "Math.random" from JavaScript, and use this to
generate a random number between 0 and 7, inclusive. This will be used
to choose which grid to fill in.

To handle the falling animation, we set the starting y offset to the
distance between this cell and the one above. If it was randomly
created, then it "should" be the maximum offset for this column. As a
simplifying hack, we can use the maximum allowable value (17 * 7 rows =
119 pixels) -- anything larger than that will overflow the signed 8-bit
offset.

Finally we animate this cell back to 0 (i.e. its regular location).