Step 29: Check for matches after dropping
After a match occurs, the cells above fall down which may create another
match. Previously we didn't check for this, which means that you could
swap cells as they were still falling.

Now we want to wait until the cells have finished falling before we give
control back to the player. We add two new states for this: "removing"
(3) and "falling" (4).

When a match occurs, instead of clearing the cells immediately, we
transition to the "removing" state, which waits for animation to finish
and then calls `$move-down` to cause the cells to fall down, and goes to
the "falling" state.

The "falling" state waits until animation finishes, and then checks for
any matched patterns. If there are any new matches, then it goes back to
the "removing" state, otherwise it goes back to the "idle" state.

Now that we have the "removing" state, we can use this for the beginning
of the game too; even we start with all cells matched then it will fill
in every cell with a random image.

To check whether animation has finished, we look at all the "time"
values for each cell. If they are all `1` then animation has finished.
As a cute trick, we can multiply all the time values; since we know that
they are all in the range [0, 1], then the result will only be `1` when
all of them are `1`.