Step 22: Change animate-cell to animate-cells
Sometimes we'll need to animate many cells with the same values
(typically resetting all the positions to 0).

Since `$animate-cell` takes a "bit" index, we can replace it with
`$animate-cells` that takes multiple bits to update.

Like `$draw-grid`, it iterates over all the bits by extracting the
least-significant `1` bit, processing that index, then clearing it and
moving on to the next bit, stopping when `$bits` is 0.