Step 23: Some small size optimizations
Now that `$animate-cells` operates on multiple bits, we can simplify
some of the animations; for example, we only want to scale up the sprite
on hover if the mouse position is not the same as the previous one. In
other words, we want `mouse-bit & ~prev-mouse-bit`.

WebAssembly doesn't have a bitwise not instruction, but it can be
implemented using the identity: `~n == n ^ -1` (i.e. xor with -1)

As another optimization, we can move the animation that resets the
"unhovered" cell after the state switch. This way it will execute
regardless of state.