Step 30: Animate match removal
Previously a removed cell would pop out of existence. It looks nicer if
they animate a little first.

We can do this by using the `$animate-cells` function on all matched
cells to shrink them down to 1 pixel.

Now that we are animating these cells, we don't want to animate the
hovered mouse all the time anymore, so we move that code up to
`$reset-prev-mouse`, which will only be executed in the "idle" and
"mouse-down" states.