Step 31: Check for game over and reset
The current game works very well, up until the point where no more moves
remain. We can detect this case using the same
`$match-all-grid-patterns` function, given 16 more patterns to check:

    .....x..
    ......xx

    ......x.
    .....x.x

    .......x
    .....xx.

    etc.

If none of these patterns match, then we know that there are no further
moves left. At this point we can reset the board so the player can start
again. We can do this by setting the match to `-1` (i.e. all cells
matched). This will clear the entire board and refill it.

Note that we now have a parameter to `$match-all-grid-patterns`, which
is the last pattern to check. This is because we only want to check the
additional patterns when looking for game over, not for regular matches.