Step 32: Display score
Finally, we keep track of score. For each cell cleared we add 1 point.
When the game ends we reset the score back to 0.

We use the `i64.popcnt` ("Population Count") instruction to count the
number of `1` bits, to determine how many points were scored for every
match.

To display the score we add 10 new sprites for the digits 0 through 9.
We add a new function `$draw-digit`, which takes an x-coordinate and a
divisor (to extract a given digit). It only draws 3 digits, so if your
score is 1000 or greater it will wrap around.