Opt: Compress data
The data is compressed using the following technique:
  for each byte in the source data:
    if the byte is <= 12 then:
      length is byte + 3
      dist is the next byte
      copy a backreference at (dist,len)
    else:
      write byte+137

The magic number 137 is a literal bias; it was calculated based on the
largest consecutive "gap" found in the uncompressed data, (i.e. the
values 137,138,...,149 are not found in the uncompressed data). Using
this bias allows us to use this range to represent a length byte and
perform just one comparison.

The magic number 3 is the minimum backreference length (since all
backreferences cost 2 bytes, it is only a savings if the copied length
is > 2 bytes). Since we know 3 is the minimum, we can encode it as 3=>0,
4=>1, etc.