Fix bug when dragging an emoji outside the grid
Thanks, Carlo P!

A few fixes are required here. First, the $get-mouse-bit function needs
to do its comparison with `i32.lt_u` so it implicitly includes a check
against negative numbers.

Next, the clicked cell shouldn't be updated when the mouse-bit is
invalid; this way it won't allow you to drag a cell outside of bounds.

Third, when the mouse is released it should reset the clicked cell.