Fix bug in inflate
When calculating the length from a code, the following table should be
used:

```
    [3, 4, 5, 6, 7, 8, 9, 10, 11, 13, 15, 17, 19, 23, 27, 31, 35, 43, 51, 59, 67, 83, 99, 115, 131, 163, 195, 227, 258]
```

As a size optimization, we use this formula instead:

```
           code <  min  =>  min + code
    min <= code <  max  =>  min + ((min + 1 + (code & min)) << extra_bits)
           code == max  =>  min + 255
```

where `min` is 3, `max` is 28 and `shift` is 2.

Before this change, the max was `29` instead. This meant that the last
entry was `259` instead of `258`.

There was a similar error in the calculation for the number of "extra
bits" to read, which caused the last entry to be 6 instead of 0.